pyproject.toml
[build-system]
requires = ["setuptools>=76", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = 'data-sitter'
version = "0.1.6"
description = "A Python library that reads data contracts and generates Pydantic models for seamless data validation."
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = 'Lázaro Pereira Candea', email = 'lazaro@candea.es'},
]
dependencies = [
    "python-dotenv==1.0.1",
    "PyYAML==6.0.2",
    "parse_type==0.6.4",
    "pydantic==2.10.5",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-cov==6.0.0",
    "pytest-mock==3.14.0",
    "twine==6.1.0",
    "build==1.2.2.post1",
]

[project.scripts]
data-sitter = "data_sitter.cli:main"
